\name{eteret}
\docType{data}
\alias{eteret}
\title{Leaf angles of a Eucalyptus tereticornis tree}
\description{
This dataset gives an example of a set of leaf angles for a Eucalyptus tree, growing 
in Richmond, NSW, Australia. 
}
\usage{eteret}
\format{A vector with 965 observations}
\references{Data courtesy of David Ellsworth. Please do not use this dataset in publications.}
\keyword{datasets}