\name{LeafAngle-package}
\alias{LeafAngle-package}
\alias{LeafAngle}
\docType{package}
\title{
Functions for fitting and displaying leaf angle distributions
}
\description{
A number of standard distributions to fit to (plant) leaf angle distributions:
a special kind of distribution where 0 < angle <90 (degrees). Based on the list of distributions
summarized by Wang et al. (2007). Uses maximum-likelihood to find parameters, and determine best fitting distributions.
}
\details{
\tabular{ll}{
Package: \tab LeafAngle\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2009-01-14\cr
License: \tab GPL \cr
LazyLoad: \tab yes\cr
}
}
\author{
Remko Duursma <remkoduursma@gmail.com>
}
\references{
Wang, W.M., Z.L. Li and H.B. Su. 2007. Comparison of leaf angle distribution functions: 
Effects on extinction coefficient and fraction of sunlit foliage. Agricultural and Forest Meteorology. 143:106-122.
}
\keyword{ package }

