\name{print.angledist}
\alias{print.angledist}
\title{ Displays an object of class angledist}
\description{
A print method for objects of class \code{angledist}.
}
\usage{
\method{print}{angledist}(x,...)
}
\arguments{
  \item{x}{ Object of class angledist }
  \item{\dots}{ None. }
}


\author{ Remko Duursma }

\seealso{ \code{\link{angledist}} }

\keyword{ misc }

