\name{print.angledistlist}
\alias{print.angledistlist}
\title{ Displays an object of class angledist}
\description{
A print method for objects of class \code{angledistlist}.
}
\usage{
\method{print}{angledistlist}(x,...)
}
\arguments{
  \item{x}{ Object of class angledistlist }
  \item{\dots}{None.}
}


\author{ Remko Duursma }

% \seealso{ }

\keyword{ misc }

