\name{summary.angledist}
\alias{summary.angledist}
\title{ Summarizes object of class angledist}
\description{
Summarize an object of class \code{angledist}.
}
\usage{
\method{summary}{angledist}(object,...)
}
\arguments{
  \item{object}{ Object of class angledist }
  \item{\dots}{None.}
}


\author{ Remko Duursma }

\seealso{ \code{\link{angledist}} }

\keyword{ misc }

